resolve PR conversations
