+++ utils/url-has-protocol.ts
